Enable Flourish FACET model configurations
Uncommented the FACET model configurations in the flourish_reference app. These configurations were commented out and not active in the system. They have been enabled now as part of a system upgrade to integrate FACET models for anxiety, depression, caregiver referrals and intimate partner violence. This allows for cleaner and more organized access to health data for these domains.